Double replaced by BigDecimal

+++ src/main/java/com/github/ccob/bittrex4j/dao/MarketOrder.java
@@ -11,26 +11,28 @@
 
 package com.github.ccob.bittrex4j.dao;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class MarketOrder {
     private int type;
-    private double quantity;
-    private double rate;
+    private BigDecimal quantity;
+    private BigDecimal rate;
     
     @JsonCreator
-    public MarketOrder(@JsonProperty("Type") int type, @JsonProperty("Quantity") double quantity, @JsonProperty("Rate") double rate) {
+    public MarketOrder(@JsonProperty("Type") int type, @JsonProperty("Quantity") BigDecimal quantity, @JsonProperty("Rate") BigDecimal rate) {
         this.quantity = quantity;
         this.rate = rate;
         this.type = type;
     }
 
-    public double getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
